feat<FE>: OpenVidu추가 및 방 생성 구현

See merge request s07-webmobile1-sub2/S07P12B204!119

+++ backend/src/main/java/com/ssafy/cheertogether/article/controller/ArticleController.java
@@ -50,7 +50,7 @@ public class ArticleController {
 
 	@GetMapping("/header/{leagueApiId}")
 	@ApiOperation(value = "헤더 별 글 목록 조회", notes = "게시판 - 헤더 별 글 목록 조회")
-	public ResponseEntity<List<ArticleResponse>> findByLeagueApiId(@PathVariable int leagueApiId) {
+	public ResponseEntity<List<ArticleResponse>> findByLeagueApiId(@PathVariable Long leagueApiId) {
 		return new ResponseEntity<>(articleService.findByHeader(leagueApiId), HttpStatus.OK);
 	}
